requirements.txt
langchain==1.2.10
langchain-ollama>=0.2.0
langchain-core==1.2.11
langchain-community>=0.4.1
langchain-text-splitters>=0.3.0
langgraph==1.0.8
langchain-mcp-adapters
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.1